+++ routes/campgrounds.js
@@ -1,23 +1,10 @@
 const express = require('express');
 const router = express.Router();
 
-const { campgroundSchema } = require('../schemas');
 const Campground = require('../models/campground');
 
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
-const { isLoggedIn } = require('../middleware');
-
-const validateCampground = (req, res, next) => {
-	const { error } = campgroundSchema.validate(req.body);
-	
-	if (error) {
-		const msg = error.details.map((el => el.message)).join(',');
-		throw new ExpressError(400, msg);
-	}
-
-	next();
-}
+const { isLoggedIn, validateCampground, isAuthor } = require('../middleware');
 
 // Campground Routes
 router.get('/', catchAsync( async (req, res) => {
@@ -53,32 +40,16 @@ router.get('/:id', catchAsync( async (req, res) => {
 	res.render('campgrounds/show', { campground });
 }));
 
-router.get('/:id/edit', isLoggedIn, catchAsync( async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync( async (req, res) => {
 	let { id } = req.params;
 	const campground = await Campground.findById(id);
-	if (!campground) {
-		req.flash('error', 'Camp not found');
-		return res.redirect('/campgrounds');
-	}
-	if (!campground.author.equals(req.user.id)) {
-		req.flash('error', 'You are not authorized to do this');
-		return res.redirect(`/campgrounds/${id}`);
-	}
 	res.render('campgrounds/edit', { campground });
 }));
 
-router.put('/:id', isLoggedIn, validateCampground, catchAsync( async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync( async (req, res) => {
 	let campground = req.body.campground;
 	let { id } = req.params;
-	const foundCampground = await Campground.findById(id);
-	if (!foundCampground) {
-		req.flash('error', 'Camp not found');
-		return res.redirect('/campgrounds');
-	}
-	if (!foundCampground.author.equals(req.user.id)) {
-		req.flash('error', 'You are not authorized to do this');
-		return res.redirect(`/campgrounds/${id}`);
-	}
+	
 	const updatedCamp = await Campground.findByIdAndUpdate(
 			id, campground, { new: true, useFindAndModify: false, runValidators: true }
 		);
@@ -86,17 +57,9 @@ router.put('/:id', isLoggedIn, validateCampground, catchAsync( async (req, res) 
 	res.redirect(`/campgrounds/${updatedCamp.id}`);
 }));
 
-router.delete('/:id', isLoggedIn, catchAsync( async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync( async (req, res) => {
 	let { id } = req.params;
-	const foundCampground = await Campground.findById(id);
-	if (!foundCampground) {
-		req.flash('error', 'Camp not found');
-		return res.redirect('/campgrounds');
-	}
-	if (!foundCampground.author.equals(req.user.id)) {
-		req.flash('error', 'You are not authorized to do this');
-		return res.redirect(`/campgrounds/${id}`);
-	}
+	
 	const deletedCamp = await Campground.findByIdAndDelete(id);
 	req.flash('success', 'Successfully deleted campground..');
 	res.redirect(`/campgrounds`);
